requirements.txt
aiogram==2.10.1
aiohttp==3.6.2
aioredis==1.3.1
aiosqlite==0.15.0
asgiref==3.2.10
async-timeout==3.0.1
asyncpg==0.21.0
attrs==20.2.0
Babel==2.8.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
Django==3.1.2
django-environ==0.4.5
gunicorn==20.0.4
h11==0.10.0
hiredis==1.1.0
idna==2.10
iso8601==0.1.13
multidict==4.7.6
pkg-resources==0.0.0
psycopg2==2.8.6
PyPika==0.42.1
pytz==2020.1
sqlparse==0.3.1
tortoise-orm==0.16.16
typing-extensions==3.7.4.3
uvicorn==0.12.1
yarl==1.6.0
